% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{draw_extracted_data}
\alias{draw_extracted_data}
\title{Get a U-MAP plot representing the pool of cells used in the clustering iteration.}
\usage{
draw_extracted_data(population, SeuratObject.init, records)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{SeuratObject.init}{a SeuratObject generated from expression.init, on which
the function RunUMAP() of Seurat has been applied already.}

\item{records}{a named list, with four data.frames: \code{cells}, \code{markers}, \code{meta} and \code{methods}.}
}
\value{
a plot.
}
\description{
Get a U-MAP plot representing the pool of cells used in the clustering iteration.
}
