% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{draw_meta_clusters}
\alias{draw_meta_clusters}
\title{Get composite U-MAP plots representing the meta-clusters predicted with the scEVE framework.}
\usage{
draw_meta_clusters(meta_clusters, data.iteration)
}
\arguments{
\item{meta_clusters}{a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{data.iteration}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{expression.init} and \code{ranking_of_genes.init}.
The two first elements correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject. and the ranking of genes generated from this matrix.
The two last elements correspond to the full scRNA-seq expression matrix and the ranking of genes generated from this matrix.}
}
\value{
a plot.
}
\description{
Get composite U-MAP plots representing the meta-clusters predicted with the scEVE framework.
}
