% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{draw_robust_clusters}
\alias{draw_robust_clusters}
\title{Get composite U-MAP plots representing the meta-clusters predicted with the scEVE framework.}
\usage{
draw_robust_clusters(robust_clusters, data.iteration)
}
\arguments{
\item{robust_clusters}{a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{data.iteration}{a named list, with two names: \code{expression} and \code{SeuratObject}.
They correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject.}
}
\value{
a plot.
}
\description{
Get composite U-MAP plots representing the meta-clusters predicted with the scEVE framework.
}
