% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{filter_conflicting_cells}
\alias{filter_conflicting_cells}
\title{Filter out conflicting cells belonging to multiple robust clusters.}
\usage{
filter_conflicting_cells(robust_clusters)
}
\arguments{
\item{robust_clusters}{a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
If a cell exists in multiple robust clusters, it is exclusively associated
to its most robust cluster. Robust clusters without cells are also removed.
}
