% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{get_associations}
\alias{get_associations}
\title{Measure the strength of the association between pairs of predicted clusters,
according to the confidence metric of the frequent itemset mining framework.
The confidence(A->C) corresponds to the proportion of cells in the cluster A, that
are also in the cluster C. Note that confidence(A->C) can differ from confidence(C->A).
Associations involving a minority of cells (i.e. confidence < 0.5) are filtered out.}
\usage{
get_associations(transaction_database)
}
\arguments{
\item{transaction_database}{a transactions object, where every cell is a transaction and every predicted
cluster is an item.}
}
\value{
a data.frame associating clusters \code{A} and \code{C} to their \code{confidence(A->C)}.
}
\description{
Measure the strength of the association between pairs of predicted clusters,
according to the confidence metric of the frequent itemset mining framework.
The confidence(A->C) corresponds to the proportion of cells in the cluster A, that
are also in the cluster C. Note that confidence(A->C) can differ from confidence(C->A).
Associations involving a minority of cells (i.e. confidence < 0.5) are filtered out.
}
