% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{get_base_clusters.default_methods}
\alias{get_base_clusters.default_methods}
\title{Apply multiple clustering methods on scRNA-seq data to predict base clusters.}
\usage{
get_base_clusters.default_methods(
  data.iteration,
  params,
  clustering_methods = c("densityCut", "monocle3", "Seurat", "SHARP")
)
}
\arguments{
\item{data.iteration}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{ranking_of_genes} and \code{expression.init}.
The three first elements correspond to the scRNA-seq expression matrix of a specific cell population,
as well as the SeuratObject and the ranking of genes generated from this matrix.
The last element corresponds to the full scRNA-seq expression matrix.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{clustering_methods}{a vector of valid clustering methods.}
}
\value{
a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.
}
\description{
Apply multiple clustering methods on scRNA-seq data to predict base clusters.
}
