% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_sceve.R
\name{get_benchmark_sceve.data}
\alias{get_benchmark_sceve.data}
\title{Using computational as well as intrinsic and extrinsic clustering metrics, measure
the performance of an instance of the scEVE framework on a dataset.}
\usage{
get_benchmark_sceve.data(data, params, method_label)
}
\arguments{
\item{data}{a named list with two elements: \code{expression.init} and \code{ground_truth}.
\code{expression.init} is a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.
\code{ground_truth} is a named factor associating cells to their cluster annotations.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{method_label}{a character.}
}
\value{
a data.frame with seven columns: \code{method}, \code{time (s)},
\code{peak_memory_usage (Mb)}, \code{ARI}, \code{NMI}, \code{Purity} and \code{SI}.
}
\description{
Using computational as well as intrinsic and extrinsic clustering metrics, measure
the performance of an instance of the scEVE framework on a dataset.
}
