% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_sceve.R
\name{get_benchmark_sceve}
\alias{get_benchmark_sceve}
\title{Using computational as well as intrinsic and extrinsic clustering metrics, measure
the performance of an instance of the scEVE framework on multiple datasets.}
\usage{
get_benchmark_sceve(datasets, params, method_label)
}
\arguments{
\item{datasets}{a vector of datasets (cf. \code{sceve::get_datasets()}).}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{method_label}{a character.}
}
\value{
a data.frame with nine columns: \code{method}, \code{dataset}, \code{time (s)},
\code{peak_memory_usage (Mb)}, \code{ARI}, \code{NMI}, \code{Purity}, \code{SI} and \code{n_cells}.
}
\description{
Using computational as well as intrinsic and extrinsic clustering metrics, measure
the performance of an instance of the scEVE framework on multiple datasets.
}
