% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_cell_ids}
\alias{get_cell_ids}
\title{Get a vector of unique cell ids according to the label of each cell.}
\usage{
get_cell_ids(labels)
}
\arguments{
\item{labels}{a vector of cell types.}
}
\value{
a vector of characters.
}
\description{
Get a vector of unique cell ids according to the label of each cell.
}
