% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_cells_of_population}
\alias{get_cells_of_population}
\title{Get every cell belonging in a predicted population.}
\usage{
get_cells_of_population(population, records.cells)
}
\arguments{
\item{population}{a character. It corresponds to a cell population predicted by scEVE (e.g. C.2).}

\item{records.cells}{a data.frame associating cells to their predicted populations.
Its rows are cells and and its columns are population. The cell values range from 0 to 1.}
}
\value{
a vector of cells.
}
\description{
Get every cell belonging in a predicted population.
}
