% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{get_cells_of_subgraph}
\alias{get_cells_of_subgraph}
\title{Get the cells at the intersection of all the clusters in a subgraph.}
\usage{
get_cells_of_subgraph(subgraph, base_clusters)
}
\arguments{
\item{subgraph}{a subgraph corresponding to strongly similar clusters.}

\item{base_clusters}{a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.}
}
\value{
a vector of cells.
}
\description{
Get the cells at the intersection of all the clusters in a subgraph.
}
