% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_characterized_clusters.markers_threshold}
\alias{get_characterized_clusters.markers_threshold}
\title{Filter out uncharacterized clusters, i.e. clusters with too little marker genes.}
\usage{
get_characterized_clusters.markers_threshold(
  clusters,
  data.iteration,
  params,
  markers_threshold = 10
)
}
\arguments{
\item{clusters}{a list where every element is a pool of cells.
The elements are named lists, with six names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers} and \code{robustness}.}

\item{data.iteration}{a named list, with two names: \code{expression} and \code{SeuratObject}.
They correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{markers_threshold}{the minimal number of marker genes expected in a characterized cluster.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with six names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers} and \code{robustness}.
}
\description{
Filter out uncharacterized clusters, i.e. clusters with too little marker genes.
}
