% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_characterized_clusters.specific_markers_threshold}
\alias{get_characterized_clusters.specific_markers_threshold}
\title{Get characterized robust clusters, i.e. robust clusters with a minimal number of specific marker genes.}
\usage{
get_characterized_clusters.specific_markers_threshold(
  robust_clusters,
  params,
  specific_markers_threshold = 22
)
}
\arguments{
\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{specific_markers_threshold}{the minimal number of specific marker genes expected in a characterized cluster.}

\item{meta_clusters}{a list where every element is a pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.
}
\description{
Get characterized robust clusters, i.e. robust clusters with a minimal number of specific marker genes.
}
