% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_iteration.R
\name{get_cluster_memberships.binary_membership}
\alias{get_cluster_memberships.binary_membership}
\title{Get a table reporting the cell composition of the characterized clusters.}
\usage{
get_cluster_memberships.binary_membership(
  characterized_clusters,
  data.iteration,
  params
)
}
\arguments{
\item{characterized_clusters}{a list where every element is a characterized pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.}

\item{data.iteration}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{expression.init} and \code{ranking_of_genes.init}.
The two first elements correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject. and the ranking of genes generated from this matrix.
The two last elements correspond to the full scRNA-seq expression matrix and the ranking of genes generated from this matrix.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}
}
\value{
a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are characterized clusters, and cell memberships are reported in the table.
}
\description{
By calling this function, the cells assigned to the leftover cluster are hard clustered,
i.e. their membership to the leftover is 1 and their membership to the other clusters is 0.
}
