% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_sceve.R
\name{get_clustering_metrics.intrinsic}
\alias{get_clustering_metrics.intrinsic}
\title{Using intrinsic metrics, measure a clustering performance.}
\usage{
get_clustering_metrics.intrinsic(expression.init, preds)
}
\arguments{
\item{expression.init}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}

\item{preds}{a named factor associating cells to their predicted clusters.}
}
\value{
a named vector of with two names: \code{Purity} and \code{SI}.
}
\description{
Using intrinsic metrics, measure a clustering performance.
}
