% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_sceve.R
\name{get_clustering_metrics}
\alias{get_clustering_metrics}
\title{Using both intrinsic and extrinsic clustering metrics, measure a clustering performance.}
\usage{
get_clustering_metrics(data, preds)
}
\arguments{
\item{data}{a named list with two elements: \code{expression.init} and \code{ground_truth}.
\code{expression.init} is a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.
\code{ground_truth} is a named factor associating cells to their cluster annotations.}

\item{preds}{a named factor associating cells to their predicted clusters.}
}
\value{
a named vector with four names: \code{ARI}, \code{NMI}, \code{Purity} and \code{SI}.
}
\description{
Extrinsic metrics compare cluster predictions to the cell annotations of the dataset.
Intrinsic metrics compare the gene expression of cells in and out of their clusters.
The \code{ARI} and the \code{NMI} are extrinsic metrics.
The \code{Purity} and the \code{SI} are intrinsic metrics.
For every metric, higher is better and the maximum value is 1.
}
