% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_sceve.R
\name{get_data.bluster}
\alias{get_data.bluster}
\title{Get a matrix usable for the functions of bluster.}
\usage{
get_data.bluster(expression.init)
}
\arguments{
\item{expression.init}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}
}
\value{
an object associating selected genes and cells to their reduced dimensions.
}
\description{
Get a matrix usable for the functions of bluster.
}
