% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data}
\alias{get_data}
\title{Get a data.frame and the ground truth of a dataset loaded with TMExplorer.}
\usage{
get_data(dataset)
}
\arguments{
\item{dataset}{a list loaded with TMExplorer::queryTME.}
}
\value{
a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.
}
\description{
Get a data.frame and the ground truth of a dataset loaded with TMExplorer.
}
