% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_iteration.R
\name{get_drawings_paths}
\alias{get_drawings_paths}
\title{Get the paths leading to each informative figure drawn during the clustering iteration.}
\usage{
get_drawings_paths(population, params)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}
}
\value{
a vector of paths.
}
\description{
Get the paths leading to each informative figure drawn during the clustering iteration.
}
