% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{get_formatted_predictions}
\alias{get_formatted_predictions}
\title{Format the cluster predictions of a method to obtain a standard output.}
\usage{
get_formatted_predictions(cells, predictions)
}
\arguments{
\item{cells}{a vector of cells names.}

\item{predictions}{a vector of cluster predictions.}
}
\value{
a named factor that associates each cell to their cluster prediction.
}
\description{
Format the cluster predictions of a method to obtain a standard output.
}
