% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_leaf_clusters}
\alias{get_leaf_clusters}
\title{Get a named vector associating each cell to its most informative cluster label.}
\usage{
get_leaf_clusters(records.cells)
}
\arguments{
\item{records.cells}{a data.frame associating cells to their predicted populations.
Its rows are cells and and its columns are population. The cell values range from 0 to 1.}
}
\value{
a named vector associating cells to their most informative cluster labels.
}
\description{
Get a named vector associating each cell to its most informative cluster label.
}
