% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_marker_genes.over_representation}
\alias{get_marker_genes.over_representation}
\title{Get marker genes characteristic of each meta-cluster by conducting an over-representation test of the expressed genes.}
\usage{
get_marker_genes.over_representation(
  meta_clusters,
  data.iteration,
  params,
  pvalue_threshold = 0.05
)
}
\arguments{
\item{meta_clusters}{a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{data.iteration}{a named list, with three names: \code{expression}, \code{SeuratObject} and \code{ranking_of_genes}.
They correspond to the scRNA-seq expression matrix of a specific cell population,
as well as the SeuratObject and the ranking of genes generated from this matrix.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{pvalue_threshold}{the pvalue threshold below which genes are considered as marker genes.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with six names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers} and \code{robustness}.
}
\description{
Get marker genes characteristic of each meta-cluster by conducting an over-representation test of the expressed genes.
}
