% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_marker_genes.seurat_findmarkers}
\alias{get_marker_genes.seurat_findmarkers}
\title{Get marker genes by calling the function \code{FindMarkers} of the Seurat package.}
\usage{
get_marker_genes.seurat_findmarkers(
  cluster,
  data.iteration,
  params,
  FC_threshold = 4,
  pvalue_threshold = 0.001
)
}
\arguments{
\item{cluster}{a named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{data.iteration}{a named list, with two names: \code{expression} and \code{SeuratObject}.
They correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{FC_threshold}{a numeric.}

\item{pvalue_threshold}{a numeric.}
}
\value{
a named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label},\code{markers} and \code{robustness}.
}
\description{
A named vector associating marker genes to their log2FC is returned.
Only genes with log2FC > 4 and p-values (corrected with Bonferonni) < 0.001 are returned.
}
