% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_max_resolution}
\alias{get_max_resolution}
\title{Get the maximum clustering resolution attained in a scEVE clustering analysis.}
\usage{
get_max_resolution(records.cells)
}
\arguments{
\item{records.cells}{a data.frame associating cells to their predicted populations.
Its rows are cells and and its columns are population. The cell values range from 0 to 1.}
}
\value{
an integer.
}
\description{
Get the maximum clustering resolution attained in a scEVE clustering analysis.
}
