% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve.R
\name{get_pending_population}
\alias{get_pending_population}
\title{Get a cell population for which no scEVE clustering iteration has been attempted.}
\usage{
get_pending_population(records)
}
\arguments{
\item{records}{a named list, with three data.frames: \code{cells}, \code{markers} and \code{meta}.}
}
\value{
a character.
}
\description{
Get a cell population for which no scEVE clustering iteration has been attempted.
}
