% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_populations_at_resolution}
\alias{get_populations_at_resolution}
\title{Get all the populations at a specific resolution.
The root population is resolution 1, and its children populations are resolution 2, etc.}
\usage{
get_populations_at_resolution(resolution, records.cells)
}
\arguments{
\item{resolution}{an integer.}

\item{records.cells}{a data.frame associating cells to their predicted populations.
Its rows are cells and and its columns are population. The cell values range from 0 to 1.}
}
\value{
a vector of population labels.
}
\description{
Get all the populations at a specific resolution.
The root population is resolution 1, and its children populations are resolution 2, etc.
}
