% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_ranking_of_genes}
\alias{get_ranking_of_genes}
\title{Rank the genes of every cell by descending order of expression, and get the resulting table.
Only expressed genes are reported, and the number of genes expressed varies across cells.
Hence, some table cells include NA values.}
\usage{
get_ranking_of_genes(expression.selected)
}
\arguments{
\item{expression.selected}{a scRNA-seq dataset of raw count expression, with selected genes.
Its rows are genes and its columns are cells.}
}
\value{
a data.frame associating cells to their genes, ranked by descending order of expression.
Its rows are ranks, its columns are cells, and genes are reported in the table.
}
\description{
Rank the genes of every cell by descending order of expression, and get the resulting table.
Only expressed genes are reported, and the number of genes expressed varies across cells.
Hence, some table cells include NA values.
}
