% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_records}
\alias{get_records}
\title{Load the records of a scEVE clustering analysis.}
\usage{
get_records(sheets_path)
}
\arguments{
\item{sheets_path}{a path where result sheets are stored.}
}
\value{
a named list, with four data.frames: \code{cells}, \code{markers}, \code{meta} and \code{methods}.
}
\description{
Load the records of a scEVE clustering analysis.
}
