% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_resolution}
\alias{get_resolution}
\title{Get the resolution level of a predicted population.}
\usage{
get_resolution(population)
}
\arguments{
\item{population}{a character. It corresponds to a cell population predicted by scEVE (e.g. C.2).}
}
\value{
an integer.
}
\description{
Get the resolution level of a predicted population.
}
