% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{get_robust_clusters}
\alias{get_robust_clusters}
\title{Extract robust clusters and a leftover cluster from a set of base clusters predicted
with multiple clustering methods.}
\usage{
get_robust_clusters(
  population,
  base_clusters,
  data.iteration,
  records,
  params,
  figures
)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{base_clusters}{a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.}

\item{data.iteration}{a named list, with two names: \code{expression} and \code{SeuratObject}.
They correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject.}

\item{records}{a named list, with four data.frames: \code{cells}, \code{markers}, \code{meta} and \code{methods}.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering iteration.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
An empty list is returned if no robust cluster was predicted.
}
