% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{get_robust_clusters.robustness_threshold}
\alias{get_robust_clusters.robustness_threshold}
\title{Apply a robustness threshold to identify robust clusters in a collection of subgraphs.}
\usage{
get_robust_clusters.robustness_threshold(
  subgraphs,
  params,
  robustness_threshold = 0.33
)
}
\arguments{
\item{subgraphs}{a list where every element is a pool of cells grouped together by multiple clustering methods.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{robustness_threshold}{a float.}
}
\value{
a list where every element is a pool of cells grouped together by multiple clustering methods.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
Apply a robustness threshold to identify robust clusters in a collection of subgraphs.
}
