% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_selected_genes.n_most_variable}
\alias{get_selected_genes.n_most_variable}
\title{Get the n most variable genes in a scRNA-seq dataset.
The variable genes are identified by calling the function FindVariableFeatures() of Seurat.}
\usage{
get_selected_genes.n_most_variable(expression, params, n_genes = 500)
}
\arguments{
\item{expression}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{n_genes}{the number of highly variable genes to sample.}
}
\value{
a vector of genes.
}
\description{
Get the n most variable genes in a scRNA-seq dataset.
The variable genes are identified by calling the function FindVariableFeatures() of Seurat.
}
