% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve.R
\name{get_SeuratObject.init}
\alias{get_SeuratObject.init}
\title{Get a SeuratObject from a scRNA-seq dataset of raw count expression, without selected genes.
This function is used once prior to a scEVE clustering analysis in order to draw the
extracted data at each clustering iteration (cf. \code{sceve::draw_extracted_data()}).}
\usage{
get_SeuratObject.init(expression.init)
}
\arguments{
\item{expression.init}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}
}
\value{
a SeuratObject, on which the function RunUMAP() of Seurat has been applied already.
}
\description{
Get a SeuratObject from a scRNA-seq dataset of raw count expression, without selected genes.
This function is used once prior to a scEVE clustering analysis in order to draw the
extracted data at each clustering iteration (cf. \code{sceve::draw_extracted_data()}).
}
