% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_SeuratObject.selected}
\alias{get_SeuratObject.selected}
\title{Get a SeuratObject from a scRNA-seq dataset of raw count expression, with selected genes.
This function is used as a pre-processing step for the Seurat and monocle3 clustering algorithms.}
\usage{
get_SeuratObject.selected(expression.selected)
}
\arguments{
\item{expression.selected}{a scRNA-seq dataset of raw count expression, with selected genes.
Its rows are genes and its columns are cells.}
}
\value{
a SeuratObject, on which the function ScaleData() of Seurat has been applied already.
}
\description{
Get a SeuratObject from a scRNA-seq dataset of raw count expression, with selected genes.
This function is used as a pre-processing step for the Seurat and monocle3 clustering algorithms.
}
