% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_specific_markers}
\alias{get_specific_markers}
\title{Get marker genes specific to each meta-cluster.}
\usage{
get_specific_markers(meta_clusters)
}
\arguments{
\item{meta_clusters}{a list where every element is a pool of cells.
The elements are named lists, with six names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers} and \code{robustness}.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.
}
\description{
Get marker genes specific to each meta-cluster.
}
