% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{get_strong_similarities}
\alias{get_strong_similarities}
\title{Get pairs of clusters that share a strong similarity.
A strong similarity is observed when two clusters share the majority of their cells,
i.e. confidence(A->C) > 0.5 & confidence(C->A) > 0.5.
It corresponds to the minimum proportion of cells shared between two clusters.}
\usage{
get_strong_similarities(associations)
}
\arguments{
\item{associations}{a data.frame associating two clusters \code{A} and \code{C}, to their \code{confidence(A->C)}.}
}
\value{
a data.frame associating two clusters \code{A} and \code{C}, to their \code{similarity}.
}
\description{
Get pairs of clusters that share a strong similarity.
A strong similarity is observed when two clusters share the majority of their cells,
i.e. confidence(A->C) > 0.5 & confidence(C->A) > 0.5.
It corresponds to the minimum proportion of cells shared between two clusters.
}
