% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{get_subgraphs}
\alias{get_subgraphs}
\title{Model a graph where every node is a base cluster predicted, and edges are strong similarities.}
\usage{
get_subgraphs(population, strong_similarities, base_clusters, params)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{strong_similarities}{a data.frame associating two clusters \code{A} and \code{C}, to their \code{similarity}.}

\item{base_clusters}{a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}
}
\value{
a list where every element is a pool of cells grouped together by multiple clustering methods.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
The graph is disjoint, and subgraphs (i.e. connected components) are extracted from it.
They will be used to identify robust clusters.
}
