% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{get_transaction_database}
\alias{get_transaction_database}
\title{"Functions used to get robust clusters, i.e. cell clusters predicted by multiple methods.}
\usage{
get_transaction_database(base_clusters)
}
\arguments{
\item{base_clusters}{a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.}
}
\value{
a transactions object, where every cell is a transaction and every predicted
cluster is an item.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{2025/01/21 @yanisaspic"
}\if{html}{\out{</div>}}

Get a transaction database from the base clusters predicted.
}
