% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{ignore_dropped_genes}
\alias{ignore_dropped_genes}
\title{Remove genes that are not expressed in a cell, from a table ranking gene expression.}
\usage{
ignore_dropped_genes(ranking_of_genes, expression.selected)
}
\arguments{
\item{ranking_of_genes}{a data.frame associating cells to their genes, ranked by descending order of expression.
Its rows are ranks, its columns are cells, and genes are reported in the table.}

\item{expression.selected}{a scRNA-seq dataset of raw count expression, with selected genes.
Its rows are genes and its columns are cells.}
}
\value{
a data.frame associating cells to their genes, ranked by descending order of expression.
Its rows are ranks, its columns are cells, and either genes or NA values are reported in the table.
}
\description{
Remove genes that are not expressed in a cell, from a table ranking gene expression.
}
