% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve.R
\name{initialize_records}
\alias{initialize_records}
\title{Get a named list, with three data.frames: \code{cells}, \code{markers} and \code{meta}.
\itemize{
\item \code{cells} associates cells to their predicted populations.
Its rows are cells, its columns are predicted populations, and cell memberships are reported in the table.
\item \code{markers} associates predicted populations to their marker genes.
Its rows are genes, its columns are predicted populations, and characterization powers are reported in the table.
\item \code{meta} associates predicted populations to generic information, including:
their \code{size}, their \code{robustness}, their \code{parent} and their \code{clustering_status}.
}}
\usage{
initialize_records(expression.init)
}
\arguments{
\item{expression.init}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}
}
\value{
a named list, with three data.frames: \code{cells}, \code{markers} and \code{meta}.
}
\description{
Get a named list, with three data.frames: \code{cells}, \code{markers} and \code{meta}.
\itemize{
\item \code{cells} associates cells to their predicted populations.
Its rows are cells, its columns are predicted populations, and cell memberships are reported in the table.
\item \code{markers} associates predicted populations to their marker genes.
Its rows are genes, its columns are predicted populations, and characterization powers are reported in the table.
\item \code{meta} associates predicted populations to generic information, including:
their \code{size}, their \code{robustness}, their \code{parent} and their \code{clustering_status}.
}
}
