% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_iteration.R
\name{merge_drawings}
\alias{merge_drawings}
\title{Merge the informative figures drawn during the clustering iteration in a single .pdf file.}
\usage{
merge_drawings(population, params)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}
}
\description{
Merge the informative figures drawn during the clustering iteration in a single .pdf file.
}
