% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve.R
\name{sceve.iteration}
\alias{sceve.iteration}
\title{Attempt to cluster a specific cell population using the scEVE algorithm.}
\usage{
sceve.iteration(
  population,
  expression.init,
  SeuratObject.init,
  records,
  params,
  figures,
  sheets
)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{expression.init}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}

\item{SeuratObject.init}{a SeuratObject generated from expression.init, on which
the function RunUMAP() of Seurat has been applied already.}

\item{records}{a named list, with three data.frames: \code{cells}, \code{markers} and \code{meta}.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering iteration.}

\item{sheets}{a boolean that indicates if the results of the clustering iteration should be saved in Excel sheets.}
}
\value{
a named list, with three data.frames: \code{cells}, \code{markers} and \code{meta}.
}
\description{
Attempt to cluster a specific cell population using the scEVE algorithm.
}
