% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve.R
\name{sceve}
\alias{sceve}
\title{Conduct a clustering analysis with the scEVE framework.}
\usage{
sceve(
  expression.init,
  params = get_default_parameters(),
  figures = TRUE,
  sheets = TRUE
)
}
\arguments{
\item{expression.init}{a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering iteration.}

\item{sheets}{a boolean that indicates if the results of the clustering iteration should be saved in Excel sheets.}
}
\value{
a named list, with two elements: \code{records} and \code{preds}.
\code{records} is a named list, with four data.frames: \code{cells}, \code{markers}, \code{meta} and \code{methods}.
\code{preds} is a named factor associating cells to their predicted clusters.
}
\description{
Conduct a clustering analysis with the scEVE framework.
}
