% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{test_over_representation}
\alias{test_over_representation}
\title{Test the over-representation of a gene in a cluster with regards to the frequency of its expression
in the cluster and in the complete pool of cells.}
\usage{
test_over_representation(
  cells_in_cluster.expressing,
  cells_in_pool.expressing,
  cells_in_cluster,
  cells_in_pool
)
}
\arguments{
\item{cells_in_cluster.expressing}{the number of cells expressing the gene in the cluster.}

\item{cells_in_pool.expressing}{the number of cells expressing the gene in the pool of cells.}

\item{cells_in_cluster}{the number of cells in the cluster.}

\item{cells_in_pool}{the number of cells in the pool of cells.}
}
\value{
the p-value of the over-representation test.
}
\description{
Test the over-representation of a gene in a cluster with regards to the frequency of its expression
in the cluster and in the complete pool of cells.
}
