% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{use_densityCut}
\alias{use_densityCut}
\title{Predict clusters with the densityCut method.}
\usage{
use_densityCut(logtpm.population, params)
}
\arguments{
\item{logtpm.population}{a scRNA-seq dataset of log-transformed transcripts per millions (tpm),
with selected genes.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}
}
\value{
a named factor that associates each cell to their cluster prediction.
}
\description{
Predict clusters with the densityCut method.
}
