% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{use_SHARP}
\alias{use_SHARP}
\title{Predict clusters with the SHARP method.}
\usage{
use_SHARP(expression.population, params)
}
\arguments{
\item{expression.population}{a scRNA-seq dataset of raw count expression, with selected genes.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}
}
\value{
a named factor that associates each cell to their cluster prediction.
}
\description{
Predict clusters with the SHARP method.
}
