% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{add_leftover_cluster}
\alias{add_leftover_cluster}
\title{Add a leftover cluster to the list of robust clusters.
It corresponds to a group of cells unassigned to any robust cluster.}
\usage{
add_leftover_cluster(population, robust_clusters, data.iteration)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{robust_clusters}{a list where every element is a pool of cells grouped together by multiple clustering methods.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{data.iteration}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{expression.init} and \code{ranking_of_genes.init}.
The two first elements correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject. and the ranking of genes generated from this matrix.
The two last elements correspond to the full scRNA-seq expression matrix and the ranking of genes generated from this matrix.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
Add a leftover cluster to the list of robust clusters.
It corresponds to a group of cells unassigned to any robust cluster.
}
