% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{generate_color_scale}
\alias{generate_color_scale}
\title{Generate a vector of colors equal to the number of identities in the sample.}
\usage{
generate_color_scale(labels)
}
\arguments{
\item{labels}{a vector of cluster labels.}
}
\value{
a named vector of colors.
}
\description{
This function is directly copied from the repository of the SCpubr package.
c.f. https://github.com/enblacar/SCpubr/blob/main/R/utils.R
}
