% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{get_base_clusters}
\alias{get_base_clusters}
\title{Get base clusters predicted with multiple clustering methods.}
\usage{
get_base_clusters(population, data.iteration, params, figures)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{data.iteration}{a named list, with two names: \code{expression} and \code{SeuratObject}.
They correspond to the scRNA-seq expression matrix of a specific cell population and its SeuratObject.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering iteration.}
}
\value{
a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.
}
\description{
Get base clusters predicted with multiple clustering methods.
}
