% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_characterized_clusters.markers_threshold}
\alias{get_characterized_clusters.markers_threshold}
\title{Check which meta-clusters are characterized, i.e. which meta-clusters have a minimum number of marker genes.
If every meta-cluster is characterized, return them. Else, return none.}
\usage{
get_characterized_clusters.markers_threshold(
  meta_clusters,
  params,
  markers_threshold = 22
)
}
\arguments{
\item{meta_clusters}{a list where every element is a pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{markers_threshold}{the minimal number of marker genes expected in a characterized cluster.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.
}
\description{
Check which meta-clusters are characterized, i.e. which meta-clusters have a minimum number of marker genes.
If every meta-cluster is characterized, return them. Else, return none.
}
