% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{get_characterized_clusters}
\alias{get_characterized_clusters}
\title{Get characterized meta-clusters. They correspond to the sub-populations predicted by the clustering iteration.}
\usage{
get_characterized_clusters(
  population,
  meta_clusters,
  data.iteration,
  params,
  figures
)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{meta_clusters}{a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{data.iteration}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{ranking_of_genes} and \code{expression.init}.
The three first elements correspond to the scRNA-seq expression matrix of a specific cell population,
as well as the SeuratObject and the ranking of genes generated from this matrix.
The last element corresponds to the full scRNA-seq expression matrix.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering iteration.}
}
\value{
a list where every element is a characterized pool of cells.
The elements are named lists, with seven names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label}, \code{markers}, \code{robustness} and \code{specific_markers}.
}
\description{
Get characterized meta-clusters. They correspond to the sub-populations predicted by the clustering iteration.
}
