% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{get_data_input}
\alias{get_data_input}
\title{Get the data input relevant to a clustering method.}
\usage{
get_data_input(method, data)
}
\arguments{
\item{method}{a clustering method used by scEVE.
Valid methods include: densityCut, monocle3, Seurat and SHARP.}

\item{data}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{logtpm} and \code{ranking_of_genes}.
They correspond to the scRNA-seq expression matrix of a specific cell population,
as well as the SeuratObject, the log-transformed TPM matrix, and the ranking of genes generated from this matrix.}
}
\value{
one of \code{data$expression}, \code{data$SeuratObject} or \code{data$logtpm}.
}
\description{
Get the data input relevant to a clustering method.
}
