% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{get_meta_clusters}
\alias{get_meta_clusters}
\title{Extract robust clusters and a leftover cluster from a set of base clusters predicted
with multiple clustering methods.}
\usage{
get_meta_clusters(population, base_clusters, data.iteration, params, figures)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{base_clusters}{a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.}

\item{data.iteration}{a named list, with four names: \code{expression}, \code{SeuratObject}, \code{ranking_of_genes} and \code{expression.init}.
The three first elements correspond to the scRNA-seq expression matrix of a specific cell population,
as well as the SeuratObject and the ranking of genes generated from this matrix.
The last element corresponds to the full scRNA-seq expression matrix.}

\item{params}{a list of parameters (cf. \code{sceve::get_default_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering iteration.}
}
\value{
a list where every element is a pool of cells.
The elements are named lists, with five names:
\code{base_clusters}, \code{cells}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
Extract robust clusters and a leftover cluster from a set of base clusters predicted
with multiple clustering methods.
}
